feat(web): add child components
