+++ apps/web/src/app/app.jsx
@@ -1,30 +1,23 @@
 import { Combobox, Dialog, Transition } from '@headlessui/react';
+import { FolderIcon, SearchIcon } from '@heroicons/react/outline';
 import concat from 'lodash/concat';
 import debounce from 'lodash/debounce';
 import shuffle from 'lodash/shuffle';
 import { Fragment, useCallback, useMemo, useState } from 'react';
-import { QueryClient, QueryClientProvider } from 'react-query';
 import { useDoughnuts } from './api';
+import ActiveOption from './components/ActiveOption';
+import Footer from './components/Footer';
+import Help from './components/Help';
+import SearchResults from './components/SearchResults';
 import { drinks } from './data';
-import Footer from './Footer';
-import Search from './Search';
-
-const queryClient = new QueryClient();
-
-export default function App() {
-  return (
-    <QueryClientProvider client={queryClient}>
-      <Home />
-    </QueryClientProvider>
-  );
-}
+import { classNames } from './helpers/utils';
 
 export const STATE = {
   EMPTY: '',
   HELP: '?'
 };
 
-function Home() {
+export default function App() {
   const { doughnuts, doughnutsLoading } = useDoughnuts();
 
   const [state, setState] = useState(STATE.EMPTY);
@@ -33,10 +26,6 @@ function Home() {
 
   const filterResults = useCallback(
     event => {
-      if (event === '') {
-        setState(event);
-      }
-
       if (event === '?') {
         setState(event);
       }
@@ -67,7 +56,7 @@ function Home() {
     [doughnuts]
   );
 
-  const debounceFilterResults = useMemo(() => debounce(filterResults, 500), [
+  const debounceFilterResults = useMemo(() => debounce(filterResults, 300), [
     filterResults
   ]);
 
@@ -125,14 +114,62 @@ function Home() {
             className="mx-auto w-full max-w-2xl max-w-screen-md transform divide-y divide-gray-500 divide-opacity-20 overflow-hidden rounded-xl bg-gray-900 shadow-2xl transition-all flex flex-col"
             onChange={item => setActiveOption(item)}
           >
-            <Search
-              onChange={handleChange}
-              {...{ activeOption, filtered, recentFiltered, state }}
-            />
-            <Footer query={state} />
+            <div className="relative">
+              <SearchIcon
+                className="pointer-events-none absolute top-3.5 left-4 h-5 w-5 text-gray-500"
+                aria-hidden="true"
+              />
+              <Combobox.Input
+                className="h-12 w-full border-0 bg-transparent pl-11 pr-4 text-white placeholder-gray-500 sm:text-sm"
+                placeholder="Search..."
+                onChange={handleChange}
+              />
+            </div>
+
+            <Combobox.Options
+              static
+              className="max-h-120 scroll-py-2 divide-y divide-gray-500 divide-opacity-20 overflow-y-auto"
+            >
+              <div
+                className={classNames(
+                  'grid divide-x divide-gray-700',
+                  activeOption ? 'grid-cols-2' : 'grid-cols-1'
+                )}
+              >
+                <div className="min-w-0 flex-auto scroll-py-4 overflow-y-auto py-2">
+                  {state === STATE.EMPTY && (
+                    <h2 className="mt-4 mb-2 px-3 text-xs font-semibold text-gray-200">
+                      Recent searches
+                    </h2>
+                  )}
+                  <SearchResults
+                    filtered={state === STATE.EMPTY ? recentFiltered : filtered}
+                  />
+                </div>
+                {activeOption != null ? (
+                  <ActiveOption option={activeOption} />
+                ) : null}
+              </div>
+            </Combobox.Options>
+
+            {state === STATE.HELP && <Help />}
+            {/* {filtered.length === 0 && <EmptyState />} */}
+            <Footer />
           </Combobox>
         </Transition.Child>
       </Dialog>
     </Transition.Root>
   );
 }
+
+const EmptyState = () => (
+  <div className="py-14 px-6 text-center sm:px-14">
+    <FolderIcon
+      className="mx-auto h-6 w-6 text-gray-500 mb-4"
+      aria-hidden="true"
+    />
+    <span className="mt-4 text-sm text-gray-200">
+      We couldn't find any projects with that term. Please try again.
+    </span>
+  </div>
+);
