--- apps/web/src/app/app.module.css
@@ -1,1 +1,0 @@
-/* Your styles goes here. */
