--- apps/web/src/app/app.spec.tsx
@@ -1,17 +1,0 @@
-import { render } from '@testing-library/react';
-
-import App from './app';
-
-describe('App', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<App />);
-
-    expect(baseElement).toBeTruthy();
-  });
-
-  it('should have a greeting as the title', () => {
-    const { getByText } = render(<App />);
-
-    expect(getByText(/Welcome web/gi)).toBeTruthy();
-  });
-});
