+++ apps/web/src/main.tsx
@@ -1,11 +1,16 @@
 import { StrictMode } from 'react';
 import * as ReactDOM from 'react-dom';
+import { QueryClient, QueryClientProvider } from 'react-query';
 import 'tailwindcss/tailwind.css';
-import App from './app/app';
+import App from './app/App';
+
+const queryClient = new QueryClient();
 
 ReactDOM.render(
   <StrictMode>
-    <App />
+    <QueryClientProvider client={queryClient}>
+      <App />
+    </QueryClientProvider>
   </StrictMode>,
   document.getElementById('root')
 );
